Click on toggle element in BulkSelect tests
