+++ packages/tsc-transform-imports/.gitignore
@@ -1,3 +1,7 @@
 index.js
 index.d.ts
 index.js.map
+directories.js
+directories.d.ts
+old-core-imports.js
+old-core-imports.d.ts
